Adjust to cissr's protocol for non-full-width stores.
For instructions like SW (and brethren), which store the low 32-bits (or W-bits)
of the register, Cissr still expects to find the full register value in the verifier
interface.  It handles the partial write in its memory model.

Since we implement our own dumb memory model, we were masking the value before the write,
and sending the masked value to the verifier.  Instead, now we send the full register
value to the verifier, which now cannot check that we masked it correctly.

Ideally, both the full value and the masked value should pass the verifier test, since both
are correct.